requirements.txt
python-telegram-bot[job-queue]
python-dotenv
supabase
google-genai
requests
Pillow
streamlit
plotly
psutil
openai
